tablecellfloat/currency: full double precision (E6)
Two precision bugs found by the new test:
- writeAttributes used QString::number(value, 'f', 2), truncating any
  non-currency float to 2 decimals on save. Switch to 'g',15 so the
  shortest representation that round-trips is emitted.
- readAttribute used value.toFloat() (32-bit) on a 64-bit double field,
  losing precision before the write path even ran. Switch to .toDouble().
  Same toFloat→toDouble fix in TablecellCurrency for consistency.

The existing 42.5 float test had to flip its office:value assertion
from "42.50" to "42.5" (the new 'g',15 output) — the Italian-style
display text:p still uses 'f',2 ("42,50") by intent.

+++ qoasis/table/tablecellcurrency.cpp
@@ -51,7 +51,7 @@ namespace qoasis::table
 	void TablecellCurrency::readAttribute(QStringRef name, QStringRef value)
 	{
 		if (name.toString() == kCellTypeAttribute) {
-			value_amount_ = value.toFloat();
+			value_amount_ = value.toDouble();
 			return;
 		}
 		if (name.toString() == kCellCurrencyAttribute) {
